+++ tests/test_monkeypatch.py
@@ -1,6 +1,7 @@
 import datetime
 
 from django.contrib import admin
+from django.template.loader import render_to_string
 from django.test import RequestFactory
 from django.urls import reverse
 
@@ -9,7 +10,7 @@ from cms.test_utils.testcases import CMSTestCase
 from djangocms_moderation import constants
 from djangocms_versioning.constants import PUBLISHED
 
-from djangocms_content_expiry.constants import CONTENT_EXPIRY_EXPIRE_FIELD_LABEL
+from djangocms_content_expiry.constants import CONTENT_EXPIRY_COMPLIANCE_FIELD_LABEL
 from djangocms_content_expiry.models import ContentExpiry
 from djangocms_content_expiry.test_utils.factories import (
     ChildModerationRequestTreeNodeFactory,
@@ -19,6 +20,7 @@ from djangocms_content_expiry.test_utils.factories import (
     UserFactory,
 )
 from djangocms_content_expiry.test_utils.polls import factories
+from djangocms_content_expiry.test_utils.polls.cms_config import PollsCMSConfig
 
 
 class ContentExpiryMonkeyPatchTestCase(CMSTestCase):
@@ -49,6 +51,7 @@ class ContentExpiryMonkeyPatchTestCase(CMSTestCase):
             collection_id=self.collection.pk,
             mr_id=self.moderation_request1.pk,
         )
+        self.versionable = PollsCMSConfig.versioning[0]
 
     def test_extended_admin_monkey_patch_list_display_expires(self):
         """
@@ -65,8 +68,8 @@ class ContentExpiryMonkeyPatchTestCase(CMSTestCase):
         list_display = version_admin.get_list_display(request)
 
         # List display field should have been added by monkeypatch
-        self.assertIn('expire', list_display)
-        self.assertEqual(CONTENT_EXPIRY_EXPIRE_FIELD_LABEL, version_admin.expire.short_description)
+        self.assertIn('compliance_number', list_display)
+        self.assertEqual(CONTENT_EXPIRY_COMPLIANCE_FIELD_LABEL, version_admin.compliance_number.short_description)
 
     def test_extended_moderation_admin_update_existing_expiry_record(self):
         """
@@ -225,3 +228,25 @@ class ContentExpiryMonkeyPatchTestCase(CMSTestCase):
         self.assertEqual(ContentExpiry.objects.count(), 2)
         self.assertEqual(ContentExpiry.objects.first().compliance_number,
                          ContentExpiry.objects.last().compliance_number)
+
+    def test_extended_versioning_admin_additional_content_settings_icon(self):
+        """
+        The additional content settings icon should be added to the version table
+        """
+        endpoint = self.get_admin_url(ContentExpiry, "change", self.content_expiry_primary.pk)
+        additional_settings_control = render_to_string(
+            'djangocms_content_expiry/admin/icons/additional_content_settings_icon.html',
+            {
+                "url": f"{endpoint}?_popup=1"
+            }
+        )
+
+        querystring = "?poll=1"
+        url = self.get_admin_url(self.versionable.version_model_proxy, "changelist") + querystring
+
+        response = self.client.get(
+            url
+        )
+
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, additional_settings_control, html=True)