+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 unreleased
 ==========
+* feat: Expiry field removed from version table and replaced with compliance number. Added additional content settings action to version table
 
 1.3.1 (2022-06-22)
 ==================